splitted class TestAppend to file test_append.py
